@@ -18,15 +18,15 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^0.1.3
-  http: 0.12.1
-  shared_preferences: ^0.5.3+4
-  flutter_dotenv: 2.1.0
+  cupertino_icons: ^1.0.2
+  http: ^0.13.0
+  shared_preferences: ^2.0.3
+  flutter_dotenv: 4.0.0-nullsafety.0
 
 dev_dependencies:
   flutter_test:
