@@ -1,10 +1,9 @@
-import 'package:flutter/cupertino.dart';
 
 class Champion {
-  int id;
-  String name;
+  late int id;
+  String? name;
 
-  Champion.min({@required this.id});
+  Champion.min({required this.id});
 
   Champion.fromJson(Map<String, dynamic> json) {
     id = int.parse(json['key']);
@@ -12,7 +11,7 @@ class Champion {
   }
 
   int compareTo(Champion other) {
-    return this.name.compareTo(other.name);
+    return this.name!.compareTo(other.name!);
   }
 
 }
