@@ -2,16 +2,16 @@ import 'package:lolinfo/Models/Champion.dart';
 import 'package:lolinfo/Networking/DDragonService.dart';
 
 class PastMatch {
-  String platformId;
-  int gameId;
-  String role;
-  int season;
-  int champId;
-  Champion champ;
-  int queue;
-  int timestamp;
-  DateTime dateTime;
-  String lane;
+  late String platformId;
+  late int gameId;
+  late String role;
+  late int season;
+  late int champId;
+  late Champion champ;
+  late int queue;
+  late int timestamp;
+  late DateTime dateTime;
+  late String lane;
 
 
   PastMatch.fromJson(Map<String, dynamic> json) {
@@ -21,7 +21,7 @@ class PastMatch {
     season = json['season'];
     champId = json['champion'];
     DDragonService.getAllChampions().then((value) {
-      champ = value[champId];
+      champ = value[champId]!;
     });
 
     queue = json['queue'];
