@@ -12,9 +12,9 @@ class RiotService {
 
 
 
-  static Future<Summoner> getSummonerInfo(String summonerName) async{
+  static Future<Summoner?> getSummonerInfo(String summonerName) async{
     String type = 'lol/summoner/v4/summoners/by-name/' + summonerName;
-    String response = await Network.get(type, null);
+    String response = await Network.get(type, Map());
     Map<String, dynamic> json = jsonDecode(response);
     if(json.containsKey('status'))
       return null;
@@ -24,15 +24,14 @@ class RiotService {
 
   static Future<Summoner> getSummonerFromPUUID(String puuid) async {
     String type = 'lol/summoner/v4/summoners/by-puuid/' + puuid;
-    String response = await Network.get(type, null);
+    String response = await Network.get(type, Map());
     return Summoner.fromJson(jsonDecode(response));
   }
 
   static Future<List<ChampionMastery>> getChampionMasteries(String summonerId) async {
-    // TODO: Return all ChampionMastery for chosen summoner
     String type = 'lol/champion-mastery/v4/champion-masteries/by-summoner/' + summonerId;
     String response = await Network.get(type, null);
-    List<ChampionMastery> _list = List();
+    List<ChampionMastery> _list = [];
     List<dynamic> json = jsonDecode(response);
     ChampionMastery curr;
     for(Map<String, dynamic> _masteryJson in json) {
@@ -44,7 +43,7 @@ class RiotService {
     return _list;
   }
 
-  static Future<Champion> getChampionInfo(int champId) async {
+  static Future<Champion?> getChampionInfo(int champId) async {
     Map<int, Champion> _list = await DDragonService.getAllChampions();
     return _list[champId];
   }
@@ -54,7 +53,7 @@ class RiotService {
     Map<String, String> queries = Map();
     queries['endIndex'] = '20';
     String response = await Network.get(type, queries);
-    List<PastMatch> _list = List();
+    List<PastMatch> _list = [];
     Map<String, dynamic> json = jsonDecode(response);
     List<dynamic> _matchList = json['matches'];
 
@@ -65,7 +64,7 @@ class RiotService {
     return _list;
   }
 
-  static Future<LiveMatch> getLiveMatch(String summonerId) async {
+  static Future<LiveMatch?> getLiveMatch(String summonerId) async {
     String type = 'lol/spectator/v4/active-games/by-summoner/' + summonerId;
     String response = await Network.get(type, null);
     Map<String, dynamic> json = jsonDecode(response);
@@ -75,5 +74,7 @@ class RiotService {
     return LiveMatch.fromJson(json);
 
   }
+
+
 
 }
